fix: отработать комментарий ревьювера + переместить папки
